Add codel.c along with its three codelets.
One of them (codel1.sk) will map to pair.sk if we run some program
transformations before we run sketch. But this is a fair bit of work.

The other (codel2.sk) needs us to extend the atom template for us to get
it working. The extended template needs 4 boolean predicates and is
available in 4preds.sk

We haven't yet made any progress on the third codelet (codel3.sk),
i.e. the one involving a square root.