Revert "Currently, we can only implement a bloom filter with k=1 hash function"
This reverts commit 7b1b90cd844993d2d848b1fa5f0541a1d050e03f.
Turns out I was wrong. This is really a bloom filter of size 256 (only
instead of all hash functions hitting the same memory, each hash
function hits its own shadow copy and the results are ored together
after the fact.). Thanks to Mihai for pointing this out. This was so
obvious in retrospect, I feel dumb!